requirements.txt
 torch>=1.10.0
torchvision>=0.11.0
torchaudio>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
librosa>=0.8.0
soundfile>=0.10.0
pydub>=0.25.0
matplotlib>=3.4.0
tqdm>=4.60.0
scipy>=1.7.0
pyyaml>=5.4.0
mediapipe>=0.8.0
transformers>=4.0.0
